remove dependency on source repo

+++ bootstrap/cmd/bootstrap/app/ksServer.go
@@ -83,7 +83,7 @@ type ksServer struct {
 	// other information about the regisry.
 	knownRegistries map[string]RegistryConfig
 
-	//GKE version to use, suggest format 1.X.Y, avoid using most recent version
+	//gkeVersionOverride allows overriding the GKE version specified in DM config. If not set the value in DM config is used.
 	// https://cloud.google.com/kubernetes-engine/docs/reference/rest/v1/projects.zones.clusters
 	gkeVersionOverride string
 
@@ -223,6 +223,9 @@ type ApplyRequest struct {
 
 	// For test: GCP service account client id
 	SAClientId string
+
+	// pass *appInfo if ks app is already on disk.
+	AppInfo *appInfo
 }
 
 var (
@@ -370,7 +373,7 @@ func (s *ksServer) CreateApp(ctx context.Context, request CreateRequest, dmDeplo
 	a, repoDir, err := s.GetApp(request.Project, request.Name, kfVersion, request.Token)
 	defer os.RemoveAll(repoDir)
 	if repoDir == "" {
-		return fmt.Errorf("Cannot clone repo from cloud source repo")
+		return fmt.Errorf("Cannot load ks app from cloud source repo")
 	}
 	envName := "default"
 	if err == nil {
@@ -472,6 +475,30 @@ func (s *ksServer) CreateApp(ctx context.Context, request CreateRequest, dmDeplo
 	if request.AutoConfigure {
 		s.autoConfigureApp(&a.App, &request.AppConfig, request.Namespace, config)
 	}
+	log.Infof("Created and initialized app at %v", a.App.Root())
+	if request.Apply {
+		components := []string{}
+		for _, comp := range request.AppConfig.Components {
+			components = append(components, comp.Name)
+		}
+		err = s.Apply(ctx, ApplyRequest{
+			Name:        request.Name,
+			KfVersion:   getRegistryVersion(request, KubeflowRegName),
+			Environment: "default",
+			Components:  components,
+			Cluster:     request.Cluster,
+			Project:     request.Project,
+			Zone:        request.Zone,
+			Token:       request.Token,
+			Email:       request.Email,
+			SAClientId:  request.SAClientId,
+			AppInfo:     a,
+		})
+		if err != nil {
+			log.Errorf("Failed to apply app: %v", err)
+			return err
+		}
+	}
 
 	if dmDeploy != nil {
 		s.UpdateDmConfig(repoDir, request.Project, request.Name, kfVersion, dmDeploy)
@@ -481,7 +508,6 @@ func (s *ksServer) CreateApp(ctx context.Context, request CreateRequest, dmDeplo
 		log.Errorf("There was a problem saving config to cloud repo; %v", err)
 		return err
 	}
-	log.Infof("Created and initialized app at %v", a.App.Root())
 	return nil
 }
 
@@ -794,7 +820,7 @@ func (s *ksServer) CloneRepoToLocal(project string, token string) (string, error
 func (s *ksServer) GetApp(project string, appName string, kfVersion string, token string) (*appInfo, string, error) {
 	repoDir, err := s.CloneRepoToLocal(project, token)
 	if err != nil {
-		return nil, "", err
+		log.Errorf("Cannot clone repo from cloud source repo")
 	}
 	appDir := path.Join(repoDir, GetRepoName(project), kfVersion, appName, KubeflowFolder)
 	_, err = s.fs.Stat(appDir)
@@ -878,12 +904,14 @@ func (s *ksServer) Apply(ctx context.Context, req ApplyRequest) error {
 		log.Errorf("Failed getting GKE cluster config: %v", err)
 		return err
 	}
-	s.projectLocks[req.Project].Lock()
-	defer s.projectLocks[req.Project].Unlock()
-	a, repoDir, err := s.GetApp(req.Project, req.Name, req.KfVersion, req.Token)
-	defer os.RemoveAll(repoDir)
-	if err != nil {
-		return err
+	targetApp := req.AppInfo
+	repoDir := ""
+	if targetApp == nil {
+		targetApp, repoDir, err = s.GetApp(req.Project, req.Name, req.KfVersion, req.Token)
+		defer os.RemoveAll(repoDir)
+		if err != nil {
+			return err
+		}
 	}
 
 	bindAccount := req.Email
@@ -938,7 +966,7 @@ func (s *ksServer) Apply(ctx context.Context, req ApplyRequest) error {
 	}
 
 	applyOptions := map[string]interface{}{
-		actions.OptionApp: a.App,
+		actions.OptionApp: targetApp.App,
 		actions.OptionClientConfig: &client.Config{
 			Overrides: &clientcmd.ConfigOverrides{},
 			Config:    clientcmd.NewDefaultClientConfig(cfg, &clientcmd.ConfigOverrides{}),
@@ -1092,29 +1120,6 @@ func finishDeployment(svc KsService, req CreateRequest, dmDeploy *deploymentmana
 		return
 	}
 
-	if req.Apply {
-		components := []string{}
-		for _, comp := range req.AppConfig.Components {
-			components = append(components, comp.Name)
-		}
-		err = svc.Apply(ctx, ApplyRequest{
-			Name:        req.Name,
-			KfVersion:   getRegistryVersion(req, KubeflowRegName),
-			Environment: "default",
-			Components:  components,
-			Cluster:     req.Cluster,
-			Project:     req.Project,
-			Zone:        req.Zone,
-			Token:       req.Token,
-			Email:       req.Email,
-			SAClientId:  req.SAClientId,
-		})
-		if err != nil {
-			log.Errorf("Failed to apply app: %v", err)
-			deploymentFailure.Inc()
-			return
-		}
-	}
 	kfDeploymentsDoneCounter.Inc()
 	kfDeploymentsDoneRaw.Inc()
 	kfDeploymentLatencies.Observe(timeSinceStart(ctx).Seconds())
